refactor: make 2 separeted folders for unit and integration tests
There is 2 commands for tests now. 
npm run test, to run unit test. npm run test:integration to run integration tests (with DB)
