+++ .github/workflows/node.js.yml
@@ -46,3 +46,17 @@ jobs:
         MAIL_SENDER_EMAIL: sender@example.com
         SECRET: fakesessionsecret
         TEST_DB_URL: postgres://postgres:postgres@localhost:5432/postgres
+    - run: npm test:integration
+      env:
+        CI: true
+        OVH_ROOM_APP_KEY: aaaaaaaaa
+        OVH_ROOM_APP_SECRET: bbbbbbbbb
+        OVH_ROOM_CONSUMER_KEY: ccccccccc
+        OVH_ROOM_ACCOUNT_NUMBER: dddddddd
+        OVH_ROOM_PHONE_NUMBER: 0033122334455
+        MAIL_PASS: fakepass
+        MAIL_SERVICE: debug
+        MAIL_USER: fakeuser
+        MAIL_SENDER_EMAIL: sender@example.com
+        SECRET: fakesessionsecret
+        TEST_DB_URL: postgres://postgres:postgres@localhost:5432/postgres
