+++ package.json
@@ -18,7 +18,8 @@
     "rollback": "knex migrate:rollback",
     "listMigrations": "knex migrate:list",
     "start": "node index.js",
-    "test": "TEST_MODE=true mocha --recursive --exit --require ./test/setup.js"
+    "test": "TEST_MODE=true mocha test/unit/*.js --recursive --exit --require ./test/setup.js",
+    "test:integration": "TEST_MODE=true mocha test/integration/*.js --recursive --exit --require ./test/setup.js"
   },
   "repository": {
     "type": "git",
